Initialisation Turrialba reorganised (as for Masatepe)

+++ BC_CAF2021_Turrialba_01-10_13-20.R
@@ -3,7 +3,7 @@
 
 ## 1. INITIALISE MCMC ##
    # nChain <- as.integer(36000)
-   nChain <- as.integer(36)
+   nChain <- as.integer(20)
    source('BC/BC_CAF2021_MCMC_init_Turrialba_01-10_13-20.R')
 
 ## 2. RUNNING THE MCMC ##
